Moving exit_data out of merc.h
