+++ code/merc.h
@@ -1736,27 +1736,7 @@ struct obj_apply_data
 };
 
 
-//
-// Exit data.
-//
-
-struct exit_data
-{
-	union
-	{
-		ROOM_INDEX_DATA *to_room;
-		short vnum;
-	} u1;
-
-	long exit_info[MAX_BITVECTOR];
-	short key;
-	char *keyword;
-	char *description;
-	EXIT_DATA *next;					// OLC
-	int orig_door;						// OLC
-	RUNE_DATA *rune;
-	bool has_rune;
-};
+#include "entity/exit_data.h"
